Move certificate extension encoding code to new crate x509::extensions.
