+++ src/rust/src/lib.rs
@@ -82,6 +82,7 @@ fn _rust(py: pyo3::Python<'_>, m: &pyo3::types::PyModule) -> pyo3::PyResult<()> 
     crate::x509::common::add_to_module(x509_mod)?;
     crate::x509::crl::add_to_module(x509_mod)?;
     crate::x509::csr::add_to_module(x509_mod)?;
+    crate::x509::extensions::add_to_module(x509_mod)?;
     crate::x509::sct::add_to_module(x509_mod)?;
     m.add_submodule(x509_mod)?;
